+++ scripts/macros/2024/spells/beaconOfHope.js
@@ -1,0 +1,6 @@
+import {beaconOfHope as beaconOfHopeLegacy} from '../../spells/beaconOfHope.js';
+export let beaconOfHope = {
+    name: beaconOfHopeLegacy.name,
+    version: beaconOfHopeLegacy.version,
+    midi: beaconOfHopeLegacy.midi
+};