+++ scripts/macros/2024/spells/bladeWard.js
@@ -1,0 +1,5 @@
+export let bladeWard = {
+    name: 'Blade Ward',
+    version: '1.1.15',
+    rules: 'modern'
+};